feat(persona): surface avatar crop widget in PersonaEditor + PersonasPanel
Extracts the inline crop widget from CharacterEditor into a reusable
AvatarCropWidget component (UI + drag/zoom/toggle logic), then wires it into
PersonaEditor's Description tab so personas get the same zoom + position
customization. The header avatar in PersonaEditor and the persona-list +
group-list avatars in PersonasPanel now apply the crop via getAvatarCropStyle.

PersonaEditor parses + serializes the persona's JSON-encoded avatarCrop on load
and save. PersonasPanel ships a small parsePersonaAvatarCrop helper that
defensively validates the JSON shape so a malformed cell can't break the list
render with NaN transforms.

Chat-side persona avatar render sites (chat bubbles, quick switcher, setup
wizards, settings drawer's persona chooser) are not yet plumbed to the new
crop in this commit — they keep displaying personas the way they did before.
A follow-up will propagate userPersonaInfo.avatarCrop through ChatArea so the
crop applies in chat too.
